requirements.txt
asgiref==3.5.0
biopython==1.79
Django==4.0.1
djangorestframework==3.13.1
numpy==1.22.1
python-decouple==3.5
sqlparse==0.4.2
tzdata==2021.5
django-cors-headers==3.11.0